update other CRUD operations for pantry

+++ src/pages/api/pantry/index.js
@@ -44,33 +44,57 @@ export default async function handler(req, res) {
     isAuthenticated(req, res, async () => {
       const { ingredient, category } = req.body
       try {
-        const newIngredient = await Pantry.create({ ingredient, category })
-        res.status(200).json(newIngredient)
+        const token = req.headers.authorization;
+        if (!token) {
+          return res.status(401).json({ message: 'Missing token' });
+        }
+
+        const decodedToken = jwt.verify(token, process.env.SECRET);
+        const userId = decodedToken.id;
+        const user = await User.findByPk(userId);
+        if (!user) {
+          return res.status(404).json({ message: 'User not found' });
+        }
+
+        const newPantryItem = await user.createPantry({ ingredient, category }); // Use the association method to create a new pantry item
+        res.status(200).json(newPantryItem);
       } catch (error) {
-        console.error(error)
-        res.status(400).json({ message: 'Failed to add ingredient to pantry' })
+        console.error(error);
+        res.status(400).json({ message: 'Failed to add ingredient to pantry' });
       }
     })
   } else if (req.method === 'DELETE') {
     // delete a single item from pantry
     isAuthenticated(req, res, async () => {
-      const { ingredient } = req.query
+      const { ingredient, category } = req.query;
       try {
-        const deletedIngredient = await Pantry.destroy({
-          where: req.query,
-        })
-        if (!deletedIngredient) {
-          return res.status(404).json({ message: 'Ingredient not found' })
+        const token = req.headers.authorization;
+        if (!token) {
+          return res.status(401).json({ message: 'Missing token' });
         }
-        res.status(200).json(`${ingredient} has been deleted`)
+
+        const decodedToken = jwt.verify(token, process.env.SECRET);
+        const userId = decodedToken.id;
+        const user = await User.findByPk(userId);
+        if (!user) {
+          return res.status(404).json({ message: 'User not found' });
+        }
+
+        const deletedIngredient = await Pantry.destroy({
+          where: { userId, ingredient, category },
+        });
+        if (!deletedIngredient) {
+          return res.status(404).json({ message: 'Ingredient not found' });
+        }
+        res.status(200).json(`{[ingredient: ${ingredient}], [category: ${category}]} has successfully been deleted`);
       } catch (error) {
-        console.error(error)
+        console.error(error);
         res
           .status(500)
-          .json({ message: 'Failed to delete ingredient from pantry' })
+          .json({ message: 'Failed to delete ingredient from pantry' });
       }
-    })
+    });
   } else {
-    res.status(400).json({ message: 'Invalid request' })
+    res.status(400).json({ message: 'Invalid request' });
   }
 }
